modularized admin model by splitting original  files into 4 separate files for better structure and maintainability
