+++ backend/middleware/auth.js
@@ -1,5 +1,5 @@
 const jwt = require('jsonwebtoken');
-const User = require('../models/User');
+const User = require('../models/user/user.model');
 
 // Middleware to authenticate JWT token
 const authenticateToken = async (req, res, next) => {
