+++ backend/routes/index.js
@@ -35,7 +35,7 @@ const AuthRoutes = require('./auth');
 // const MedicalRecordRoutes = require('../modules/medicalRecord/medicalRecord.route');
 const LabTestRoutes = require('../models/labTest/labTest.route');
 // const PaymentRoutes = require('../modules/payment/payment.route');
-// const AdminRoutes = require('../modules/admin/admin.route');
+const AdminRoutes = require('../models/admin/admin.route');
 // const ChatRoutes = require('../modules/chat/chat.route');
 // const AnalyticsRoutes = require('../modules/analytics/analytics.route');
 // const HealthDataRoutes = require('../modules/healthData/healthData.route');
@@ -51,7 +51,7 @@ router.use('/auth', AuthRoutes);
 // router.use('/medical-records', MedicalRecordRoutes);
 router.use('/lab-tests', LabTestRoutes);
 // router.use('/payments', PaymentRoutes);
-// router.use('/admin', AdminRoutes);
+router.use('/admin', AdminRoutes);
 // router.use('/chat', ChatRoutes);
 // router.use('/analytics', AnalyticsRoutes);
 // router.use('/health', HealthDataRoutes);
