+++ backend/socket/chatSocket.js
@@ -1,5 +1,5 @@
 const jwt = require('jsonwebtoken');
-const User = require('../models/User');
+const User = require('../models/user/user.model');
 const Appointment = require('../models/Appointment');
 
 module.exports = (io) => {
